GUI update
Finished gui and cleaned up file structure.
