+++ code/main.py
@@ -7,7 +7,7 @@ menace = menace.MENACE('MENACE0')
 def loop():
     pos = menace.gen_move(app.output_board)
     while app.update(pos, 'X') == -1:
-        pos = menace.gen_move(app.output_board)
+        pos = menace.random(app.output_board)
 
     # board = [[0 for _ in range(3)] for _ in range(3)]
     # for (x, y), s in app.output_board():
@@ -20,7 +20,7 @@ def loop():
     # print()
 
     if app.is_win():
-        print()
+        # print()
         return
 
     pos = menace.gen_move(app.output_board)
@@ -37,7 +37,7 @@ def loop():
     #     print()
 
     if app.is_win():
-        print()
+        # print()
         return
 
     # while app.update([int(i) for i in pos], 'O') == -1:
@@ -45,17 +45,29 @@ def loop():
     #     while len(pos) != 2:
     #         pos = input('> ').split(',')
 
-    print()
+    # print()
 
+
+n = 0
 
 for set in range(50):
+    n += 1
+    print(n)
     for round in range(1000):
         app = board.Board()
+        '''
+        pos = menace.gen_move(app.output_board)
+        while app.update(pos, 'X') == -1:
+            pos = menace.random(app.output_board)
+        '''
+
         while not app.is_win():
             loop()
-
+        '''
         print('Winner:', app.is_win())
         print(set * 1000 + round)
+        print(menace.hist)
+        '''
         menace.update(app.is_win())
     menace.save()
 print('Done!')
